# openapi.yaml
...
swagger: "2.0"
info:
  version: "1.0.0"
  title: "IDUNA Platform IAM + Governance API"
  description: |
    IDUNA is the FARTHQ central trust authority. Consumer services trust IDUNA-issued
    JWTs, not external OAuth providers directly. This contract documents the implemented
    IAM surface: Google ID token exchange, agent machine-to-machine credential exchange,
    JWKS publication, unified identity entitlements, and Apples golden documentation log
    streaming.
schemes:
  - https
  - http
host: "api.example.com"
basePath: "/"
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Bearer <IDUNA-issued JWT>"
...
paths:
...
  /admin/kanban:
    get:
      tags: [Kanban]
      summary: "Kanban board HTML — 3 columns (Backlog/Priority/Cruise), drag-and-drop"
      description: >
        Prioritization layer over EMILY/BACKLOG.md — backlog item ids (e.g. "S202-27") stay
        the authoritative reference in BACKLOG.md itself; this only tracks which of the 3
        columns a card sits in. Cookie-authenticated (same session as the rest of /admin/*).
      security:
        - BearerAuth: []
      produces:
        - text/html
      responses:
        200:
          description: "Kanban board"
        403:
          description: "Missing iduna.admin permission"
...
  /api/v1/kanban/cards/{id}/comments:
    parameters:
      - name: id
        in: path
        type: integer
        required: true
    get:
      tags: [Kanban]
      summary: "List a card's comments, oldest first (each carries the author's login)"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "[{id, card_id, author, author_sub, body, created_at}]"
        404:
          description: "Card not found"
    post:
      tags: [Kanban]
      summary: "Comment on a card -- author is taken from the caller's token (agent_name, else email, else sub)"
      security:
        - BearerAuth: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              body:
                type: string
      responses:
        201:
          description: "Stored comment"
        400:
          description: "Empty or over-long body"
        404:
          description: "Card not found"
...